requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
marshmallow==3.20.1
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
python-dotenv==1.0.0